Move deadzone & inversion to react-gamepad, organize controls, format
