+++ src/components/help/GamepadTable.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import './GamepadTable.css'
-import {GamepadControls} from '../../constants/controls.js'
+import {GamepadControls} from '../../constants/controls/gamepadControls.js'
 import {useSelector, useStore} from 'react-redux'
 import type {RootState} from '../../store/store.js'
 import {isAxis, isButton, type GamepadNames} from '../../constants/gamepadConstants.js'
